requirements.txt
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
gunicorn==22.0.0
dj-database-url==2.2.0
psycopg2==2.9.9